Store multiple parsers per host and use the best fit
This assumes that the "best fit" is the parser with most fields that
could be parsed and that, in total, finds the most elements on the page.

These are saved per host so that multiple pages for a host can be used
to build up a better parsing model of the site. Only one is saved per
page path because it is assumed to be the best fit.